requirements.txt
actionlib==1.12.1
addict==2.4.0
angles==1.9.12
anyio==3.3.0
argcomplete==1.12.3
argon2-cffi==20.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bondpy==1.8.5
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
controller-manager==0.18.3
controller-manager-msgs==0.18.3
cv-bridge==1.13.0
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
deprecation==2.1.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
dynamic-reconfigure==1.6.3
einops==0.3.0
entrypoints==0.3
evo==1.13.5
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
idna==3.2
image-geometry==1.13.0
interactive-markers==1.11.5
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
joint-state-publisher==1.12.15
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-packaging==0.10.4
jupyter-server==1.10.2
jupyterlab==3.1.1
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
kdl-parser-py==1.13.1
kiwisolver==1.3.1
laser-geometry==1.6.7
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
message-filters==1.14.11
MinkowskiEngine==0.5.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random
mkl-service==2.4.0
natsort==7.1.1
nbclassic==0.3.1
nbconvert==4.3.0
nbformat==5.1.3
notebook==6.0.3
numpy
olefile==0.46
open3d==0.13.0
packaging==21.0
pandas==1.3.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-qt-binding==0.4.3
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-unixsocket==0.2.0
resource-retriever==1.12.6
rosbag==1.14.11
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.11
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rosnode==1.14.11
rosparam==1.14.11
rospy==1.14.11
rosservice==1.14.11
rostest==1.14.11
rostopic==1.14.11
rosunit==1.14.9
roswtf==1.14.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.17
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.7.1
sensor-msgs==1.12.8
six
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
sniffio==1.2.0
terminado==0.10.1
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
threadpoolctl==2.2.0
tomlkit==0.7.2
topic-tools==1.14.11
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typing-extensions
urdfdom-py==0.4.5
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==1.1.0
widgetsnbextension==3.5.1
xacro==1.13.11
